# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test JARIVIS - IQBAL IKHLASUL AMAL",
        "contact": {
            "name": "Iqbal Ikhlasul Amal",
            "email": "iqbaliamal@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pertemuan/list": {
            "get": {
                "tags": [
                    "Pertemuan"
                ],
                "summary": "Retrieve all pertemuans",
                "operationId": "59475e460a9a8001338a1ed954ac63df",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pertemuan/store": {
            "post": {
                "tags": [
                    "Pertemuan"
                ],
                "summary": "Create a new pertemuan",
                "operationId": "e09c637e6905ccac070f69f78de986b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "deleted_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pertemuan/detail/{id}": {
            "get": {
                "tags": [
                    "Pertemuan"
                ],
                "summary": "Retrieve a pertemuan",
                "operationId": "a62a505a983194feeb0539070e5a01ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pertemuan ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pertemuan/update/{id}": {
            "put": {
                "tags": [
                    "Pertemuan"
                ],
                "summary": "Update a pertemuan",
                "operationId": "82e2ea2fdcacd7b7e7018de8d40245c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pertemuan ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "deleted_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pertemuan/delete/{id}": {
            "delete": {
                "tags": [
                    "Pertemuan"
                ],
                "summary": "Delete a pertemuan",
                "operationId": "010f8c4e53b0bead9178d001ea48d072",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pertemuan ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}